# task-manager/backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task Manager API",
    "description": "API for managing tasks and task-related entities",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Main API"
    },
    {
      "url": "/api/v1",
      "description": "Task Management API v1"
    }
  ],
...
  "paths": {
...
    "/v1/jtbd": {
      "get": {
        "tags": ["Jobs To Be Done"],
        "summary": "Get all jobs to be done",
        "responses": {
          "200": {
            "description": "List of jobs to be done",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JTBD"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Jobs To Be Done"],
        "summary": "Create a new job to be done",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JTBDInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job to be done created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JTBD"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/jtbd/{jtbdId}": {
      "get": {
        "tags": ["Jobs To Be Done"],
        "summary": "Get job to be done by ID",
        "parameters": [
          {
            "name": "jtbdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job to be done details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JTBD"
                }
              }
            }
          },
          "404": {
            "description": "Job to be done not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Jobs To Be Done"],
        "summary": "Update job to be done",
        "parameters": [
          {
            "name": "jtbdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JTBDInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job to be done updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JTBD"
                }
              }
            }
          },
          "404": {
            "description": "Job to be done not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Jobs To Be Done"],
        "summary": "Delete job to be done",
        "parameters": [
          {
            "name": "jtbdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Job to be done deleted successfully"
          },
          "404": {
            "description": "Job to be done not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "JTBD": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "epicId": {
            "type": "string"
          },
          "statement": {
            "type": "string"
          },
          "userType": {
            "type": "string"
          },
          "contextualDetails": {
            "type": "string"
          },
          "motivations": {
            "type": "string"
          },
          "expectedOutcome": {
            "type": "string"
          },
          "priority": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          },
          "status": {
            "type": "string",
            "enum": ["identified", "validated", "in_progress", "completed"]
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "JTBDInput": {
        "type": "object",
        "required": ["statement", "userType", "expectedOutcome"],
        "properties": {
          "epicId": {
            "type": "string"
          },
          "statement": {
            "type": "string"
          },
          "userType": {
            "type": "string"
          },
          "contextualDetails": {
            "type": "string"
          },
          "motivations": {
            "type": "string"
          },
          "expectedOutcome": {
            "type": "string"
          },
          "priority": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          },
          "status": {
            "type": "string",
            "enum": ["identified", "validated", "in_progress", "completed"],
            "default": "identified"
          },
          "createdBy": {
            "type": "string"
          }
        }
      },
...
    }
  }
} 